feature(Validator): Добавлена перегрузка метода для разных типов значений.

+++ src/Programming/Programming/Model/Classes/Film.cs
@@ -41,7 +41,7 @@ class Film
         get { return duration; }
         set
         {
-            duration = (int)Validator.AssertOnPositiveValue((float)value, "Duration");
+            duration = Validator.AssertOnPositiveValue(value, "Duration");
         }
     }
